LayoutTests/platform/mac/accessibility/element-busy-changed.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
<title>Element Busy Changed</title>
</head>
<!-- Busy state begins false. -->
<body aria-busy="false" id="body">

<p id="description"></p>
<div id="console"></div>

<script>
    description("This tests that AXElementBusyChanged is posted correctly.");
    
    var body = document.getElementById("body");
    var elementBusyChangedCount = 0;
    function elementBusyChangedCallback(notification) {
        if (notification == "AXElementBusyChanged" && ++elementBusyChangedCount == 2) {
            shouldBe("elementBusyChangedCount", "2");
            
            wasPostTestScriptParsed = true;
            finishJSTest();
            
            if (window.testRunner)
                testRunner.notifyDone();
        }
    }
    
    if (window.testRunner)
        testRunner.waitUntilDone();
    
    if (window.accessibilityController) {
        accessibilityController.accessibleElementById("body").addNotificationListener(elementBusyChangedCallback);
        
        // Toggle through both busy state transitions.
        body.setAttribute("aria-busy", "true");
        body.setAttribute("aria-busy", "false");
    }
</script>

</body>
</html>
