LayoutTests/platform/mac/accessibility/element-busy-changed-expected.txt

This tests that AXElementBusyChanged is posted correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS elementBusyChangedCount is 2
PASS successfullyParsed is true

TEST COMPLETE
